Convert API

+++ src/main/java/com/vilai/quote/controller/QuoteController.java
@@ -19,20 +19,20 @@ public class QuoteController {
 @PostMapping("/")
 public QuoteResponse createQuote(@RequestBody QuoteCreateRequest quoteRequest) {
 	
-	QuoteResponse quoteResponse = new QuoteResponse();
-	
-	quoteResponse.setNumber(101);
-	quoteResponse.setCustomerId(quoteRequest.getCustomerId());
-	quoteResponse.setContractStartDate(quoteRequest.getContractStartDate());
-	quoteResponse.setContractEndDate(quoteRequest.getContractEndDate());
-	
-	long tcv = quoteRequest.getItems().stream().mapToInt(Item::getPrice).sum();
-	quoteResponse.setTcv(tcv);
-	
-	quoteResponse.setStatus("Open");
-	
-	quoteResponse.setItems(quoteRequest.getItems());
-	
-	return quoteResponse;
-}
+		QuoteResponse quoteResponse = new QuoteResponse();
+		
+		quoteResponse.setNumber(101);
+		quoteResponse.setCustomerId(quoteRequest.getCustomerId());
+		quoteResponse.setContractStartDate(quoteRequest.getContractStartDate());
+		quoteResponse.setContractEndDate(quoteRequest.getContractEndDate());
+		
+		long tcv = quoteRequest.getItems().stream().mapToLong(Item::getPrice).sum();
+		quoteResponse.setTcv(tcv);
+		
+		quoteResponse.setStatus("Open");
+		
+		quoteResponse.setItems(quoteRequest.getItems());
+		
+		return quoteResponse;
+	}
 }
